django/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
networkx==3.3
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sklearn==0.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.0
tzdata==2024.1